Add question 7 and add minor changes in questions 13 and 15 for exercise list 4

+++ listas/l4/13.cpp
@@ -19,7 +19,7 @@ int main(){
     int acc_time = jobs[0].first;
     int sum = acc_time * jobs[0].second;
 
-    for(int i=1;i<jobs.size();i++){
+    for(int i=1;i<(int)jobs.size();i++){
         acc_time += jobs[i].first;
         sum += (jobs[i].second * acc_time);
     }
